% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts_by_status.R
\name{accounts_by_status}
\alias{accounts_by_status}
\title{Analyze account activity status over time periods}
\usage{
accounts_by_status(account_id, order_date, by = "month", with_counts = FALSE)
}
\arguments{
\item{account_id}{A vector of account IDs}

\item{order_date}{A vector of order dates corresponding to the account IDs}

\item{by}{The time period to group by. Defaults to "month", but anything
supported as a \code{unit} for \link[lubridate:round_date]{lubridate::floor_date} and  \code{by} for
\link{seq.Date} is an option, e.g. "week", "quarter", "2 months" etc.}

\item{with_counts}{Logical, if TRUE, include counts of accounts in each
status category}
}
\value{
A data frame with columns for period start and end dates, lists of
account IDs in each status category, and optionally counts of accounts in
each category

The function returns a data frame with the following columns:
\itemize{
\item \code{period_start} (date): The start date of the period.
\item \code{period_end} (date) : The end date of the period.
\item \code{active} (list of character): Accounts with orders in the current
period.
\item \code{new} (list of character): Accounts with their first-ever order in
the current period.
\item \code{returning} (list of character): Accounts with orders in both the
current and prior periods.
\item \code{temporarily_lost} (list of character): Accounts with orders in
the prior period but not in the current period, yet have orders in future
periods.
\item \code{terminally_lost} (list of character): Accounts with orders in the
prior period but not in the current period, and no orders in future periods.
\item \code{regained} (list of character): Accounts with orders in the current period, no orders in
the prior period, but had orders in earlier periods. i.e. an account that
comes back after being "temporarily lost"
\item \code{cumulative} (list of character): All accounts that have ever had an order
up to and including the current period.
}

if \code{with_counts} is TRUE, additional columns are included:
\itemize{
\item \code{n_active} (int): Count of active accounts in the current period.
\item \code{n_new} (int): Count of new accounts in the current period.
\item \code{n_returning} (int): Count of returning accounts in the current period.
\item \code{n_temporarily_lost} (int): Count of temporarily lost accounts in the
current period.
\item \code{n_terminally_lost} (int): Count of terminally lost accounts in the
current period.
\item \code{n_regained} (int): Count of regained accounts in the current period.
\item \code{n_cumulative} (int): Count of all accounts that have ever had an order
up to and including the current period.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function categorizes accounts into various statuses (new, returning,
temporarily lost, regained, terminally lost and cumulative) over specified
time periods (monthly or quarterly). This is useful for understanding
customer retention and churn.
}
\examples{
set.seed(1234)
n <- 50
dates <- seq(as.Date("2022-01-01"), as.Date("2022-12-31"), by = "day")
orders <- data.frame(
 account_id = sample(letters[1:10], n, replace = TRUE),
 order_date = sample(dates, n, replace = TRUE)
)

accounts_by_status(orders$account_id, orders$order_date, with_counts = TRUE)
}
