% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dt.R
\name{auto_dt}
\alias{auto_dt}
\title{creates a datatable with some automatic formatting}
\usage{
auto_dt(data, pct_digits = 1, currency_digits = 0, numeric_digits = 0, ...)
}
\arguments{
\item{data}{a data frame to be displayed as a datatable}

\item{pct_digits}{number of digits to display for percentage columns}

\item{currency_digits}{number of digits to display for currency columns}

\item{numeric_digits}{number of digits to display for numeric columns}

\item{...}{additional arguments passed to \code{\link[=guess_col_fmts]{guess_col_fmts()}}}
}
\value{
a datatable with formatted columns, filter at top, and no rownames
}
\description{
This function creates a datatable with automatic formatting for percentage,
currency, and numeric columns, using \code{\link[=guess_col_fmts]{guess_col_fmts()}} to determine which
columns are percentages, currency, or numeric.
}
