% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizdays_between.R
\name{bizdays_between}
\alias{bizdays_between}
\title{Calculate Business Days Between Two Dates in a Given QuantLib Calendar}
\usage{
bizdays_between(from, to, calendar, include_first = TRUE, include_last = TRUE)
}
\arguments{
\item{from, to}{start and end dates. Date object or something coercible
with \code{\link[=as.Date]{as.Date()}}.}

\item{calendar}{(character) A QuantLib calendar id (the vector \link[qlcal:calendars]{qlcal::calendars} lists all valid options).}

\item{include_first, include_last}{(logical) Whether to include the first and last dates in the count. Defaults to \code{TRUE}.}
}
\value{
An integer representing the number of business days between the two dates, according to the specified calendar.
}
\description{
Wrapper around \code{\link[qlcal:businessDaysBetween]{qlcal::businessDaysBetween()}} that allows the use of a
specified calendar without making persistent changes to which calendar is in
use globally.
}
\examples{
bizdays_between("2025-07-01", "2025-07-15", "UnitedStates")
bizdays_between("2025-07-01", "2025-07-15", "UnitedKingdom")
}
\seealso{
\code{\link[qlcal:calendars]{qlcal::calendars()}}, \code{\link[qlcal:businessDaysBetween]{qlcal::businessDaysBetween()}}
}
