% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter_breaks.R
\name{calc_quarter_breaks}
\alias{calc_quarter_breaks}
\title{Calculate quarterly breaks for dates}
\usage{
calc_quarter_breaks(dates)
}
\arguments{
\item{dates}{A vector of dates, or something coercible with
\link[base:as.Date]{base::as.Date}}
}
\value{
A vector of dates representing the start of each quarter
within the range of \code{dates}.
}
\description{
This function calculates quarterly breaks for a given range of dates.
}
\details{
These breaks are floor dates for the quarter, so
\code{min(dates) >= min(calc_quarter_breaks(dates))}
and \code{max(dates) <= max(calc_quarter_breaks(dates))}.
}
