% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter_breaks.R
\name{calc_year_breaks}
\alias{calc_year_breaks}
\title{Calculate yearly breaks for dates}
\usage{
calc_year_breaks(dates)
}
\arguments{
\item{dates}{A vector of dates, or something coercible with
\link[base:as.Date]{base::as.Date}}
}
\value{
A vector of dates representing the start of each year
within the range of \code{dates}.
}
\description{
This function calculates yearly breaks for a given range of dates.
}
\details{
These breaks are floor dates for the year, so
\code{min(dates) >= min(calc_year_breaks(dates))}
and \code{max(dates) <= max(calc_yearr_breaks(dates))}.
}
