% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dt.R
\name{guess_col_fmts}
\alias{guess_col_fmts}
\title{Guesses appropriate format class for each numeric column in a data frame}
\usage{
guess_col_fmts(
  data,
  pct_flags = c("frac", "pct", "percent"),
  curr_flags = c("revenue", "asp", "cogs")
)
}
\arguments{
\item{data}{a data frame}

\item{pct_flags}{a character vector of flags to identify percentage columns}

\item{curr_flags}{a character vector of flags to identify currency columns}
}
\value{
a list with three character vectors: numeric, pct, currency,
with the names of the columns in each class
}
\description{
Using column names, this function classifies numeric columns into
three classes: percentage, currency, and numeric. It uses the presence of
specific flags in the column names to determine the class of each column.
Numeric columns whose names don't contain any \code{pct_flags} or \code{curr_flags}
are classified as numeric.
}
