% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bizday.R
\name{is_bizday}
\alias{is_bizday}
\title{Check if Date is a Business Day in a Given Calendar}
\usage{
is_bizday(date, calendar)
}
\arguments{
\item{date}{A vector of dates (Date object or coercible with \code{\link[=as.Date]{as.Date()}}).}

\item{calendar}{(character) A QuantLib calendar id (the vector \link[qlcal:calendars]{qlcal::calendars} lists all valid options).}
}
\value{
A logical vector indicating whether each date is a business day in the specified calendar.
}
\description{
This function checks if the provided date(s) are business days according to the specified QuantLib calendar.
}
\examples{
# valid calendars
qlcal::calendars
# July 4, 2024 is a US holiday, but not a UK holiday
is_bizday(c("2024-07-04", "2024-07-05"), "UnitedStates")
is_bizday(c("2024-07-04", "2024-07-05"), "UnitedKingdom")
}
\seealso{
\code{\link[qlcal:calendars]{qlcal::calendars()}}, \code{\link[qlcal:isBusinessDay]{qlcal::isBusinessDay()}}
}
