% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter_breaks.R
\name{label_quarters_short}
\alias{label_quarters_short}
\title{Generate labels for quarters with year}
\usage{
label_quarters_short()
}
\value{
A function that takes a vector of dates and returns labels in
the format "Qx\\nYYYY" or "Qn"
}
\description{
This function generates labels for quarters in a short format,
showing the quarter and year only when it changes from the previous label,
similar to \code{\link[scales:label_date]{scales::label_date_short()}}.
}
\details{
This should generally be used in conjunction with breaks that are
the dates of the start of a quarter, e.g. as from
\code{\link[=breaks_quarters]{breaks_quarters()}}.
}
