% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_calendars.R
\name{load_calendars}
\alias{load_calendars}
\title{Load QuantLib Calendars to cover the years spanning given dates}
\usage{
load_calendars(dates, calendars, financial = FALSE, quiet = TRUE)
}
\arguments{
\item{dates}{A vector of dates (Date objects or coercible with \code{\link[=as.Date]{as.Date()}})}

\item{calendars}{A character vector of RQuantLib calendar names
(see \code{\link[bizdays:other-calendars]{bizdays::load_quantlib_calendars()}}). Don't include the "QuantLib/"
prefix.}

\item{financial}{see \code{\link[bizdays:other-calendars]{bizdays::load_quantlib_calendars()}}. Defaults to
\code{FALSE}, which has the result that the ending business day is considered
when counting elapsed bizdays between two dates, which yields the results
we're looking for here (i.e. inclusive counting). Note this is the opposite
of the default behavior of \code{\link[bizdays:other-calendars]{bizdays::load_quantlib_calendars()}}.}

\item{quiet}{Logical indicating whether to suppress messages during loading
(default: TRUE)}
}
\value{
Invisibly returns NULL. Calendars are loaded.
}
\description{
Convenience wrapper around \code{\link[bizdays:other-calendars]{bizdays::load_quantlib_calendars()}} that
calculates the appropriate \code{from} and \code{to} arguments to cover the entire
years spanned by a vector of dates.
}
