% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrutils-package.R
\docType{package}
\name{mcrutils-package}
\alias{mcrutils}
\alias{mcrutils-package}
\title{mcrutils: Miscellaneous Utilities for Data Cleaning, Analysis, and Reporting}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Michael Caselli \email{jane@example.com}

}
\keyword{internal}
