% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_day_of_week.R
\name{most_recent_x_day}
\alias{most_recent_x_day}
\alias{most_recent_monday}
\alias{most_recent_sunday}
\title{Get the date of the prior Sunday or Monday}
\usage{
most_recent_monday(as_of = NULL)

most_recent_sunday(as_of = NULL)
}
\arguments{
\item{as_of}{A date object or a string representing the date to evaluate
relative to. Defaults to NULL, which uses the current date.}
}
\value{
A Date object representing the most recent Sunday or Monday
}
\description{
This function returns the date of the most recent Sunday or Monday that
is on or before the current date or a specified reference date.
}
\examples{
most_recent_monday() # Returns the most recent Monday from today
# Returns c("2025-12-29", "2026-01-05")
most_recent_monday(c("2026-01-04", "2026-01-05"))
most_recent_sunday("2024-06-15") # Returns "2024-06-09"
}
