% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{normalize_logicals}
\alias{normalize_logicals}
\title{Normalize character or factor columns containing logical data to logical type}
\usage{
normalize_logicals(data, quiet = FALSE)
}
\arguments{
\item{data}{A data frame or tibble}

\item{quiet}{If \code{TRUE}, suppresses messages about the columns being converted}
}
\value{
A data frame or tibble with the specified columns converted to logical type
}
\description{
Normalize character or factor columns containing logical data to logical type
}
\details{
uses \code{\link[=detect_hidden_logicals]{detect_hidden_logicals()}} to find columns that are character or factor but contain logical values.
Converts these columns to logical type using \code{\link[base:logical]{base::as.logical()}}.
}
