% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accounts_by_status.R
\name{plot_accounts_by_status}
\alias{plot_accounts_by_status}
\title{Plot Accounts by Status Over Time}
\usage{
plot_accounts_by_status(
  data,
  account_id,
  order_date,
  by = "month",
  force_final_period_complete = FALSE,
  include_cumulative = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble of order information containing at least
account IDs and order dates}

\item{account_id, order_date}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
columns in \code{data} corresponding to an account identifier and order dates,
erspectively}

\item{by}{The time period resolution. Defaults to "month", but anything
supported as a \code{unit} argument for \link[lubridate:round_date]{lubridate::floor_date} and \code{by} for
\link{seq.Date} is an option, e.g. "week", "quarter", "2 months" etc.}

\item{force_final_period_complete}{Logical, if TRUE, treat the final period
as complete even if it may not be. This forces the final period to be
displayed with solid lines, even if the period includes dates greater than
the final \code{order_date} in \code{data}.}

\item{include_cumulative}{Logical, if TRUE, include the cumulative account
counts in the plot. Defaults to TRUE.}
}
\value{
A ggplot2 object representing the line plot of count of each account
status over time (active, new, returning, temporarily lost, terminally
lost, regained, and optionally cumulative).  If
\code{force_final_period_complete} is FALSE, the final period may be displayed
with dashed lines to indicate it may be incomplete.
}
\description{
This function generates a line plot visualizing the counts of accounts by
their status over time. It uses the \code{\link[=accounts_by_status]{accounts_by_status()}} function to
categorize accounts and then creates a plot using \code{ggplot2}.
}
\examples{
example_sales |>
  plot_accounts_by_status(account_id, order_date, by = "quarter")
}
