% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytd_py.R
\name{py_dates}
\alias{py_dates}
\title{Calculate previous year dates}
\usage{
py_dates(dates, ...)
}
\arguments{
\item{dates}{a date-time object of class POSIXlt, POSIXct or Date.}

\item{...}{
  Arguments passed on to \code{\link[lubridate:mplus]{lubridate::add_with_rollback}}
  \describe{
    \item{\code{roll_to_first}}{rollback to the first day of the month instead of the
last day of the previous month (passed to \code{\link[lubridate:rollback]{rollback()}})}
    \item{\code{preserve_hms}}{retains the same hour, minute, and second information? If
FALSE, the new date will be at 00:00:00 (passed to \code{\link[lubridate:rollback]{rollback()}})}
  }}
}
\value{
A vector of dates one year prior to the input dates.
Any fictitious resulting dates (e.g. the result of subtracting a year
from a leap day) are rolled back to the prior valid date.
}
\description{
Calculates the previous year dates
by subtracting one year from the given dates.
}
