% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytd_py.R
\name{ytd_bounds}
\alias{ytd_bounds}
\title{Find the start and end dates of the year-to-date period}
\usage{
ytd_bounds(dates, rollback_partial_month = FALSE)
}
\arguments{
\item{dates}{a vector of dates, or something coercible with
\code{\link[lubridate:as_date]{lubridate::as_date()}}}

\item{rollback_partial_month}{If TRUE, and the maximum date in \code{dates}
is not an end-of-month date, the end date will be rolled back to the
last day of the previous month. If FALSE the end date will be the maximum
date in \code{dates}.}
}
\value{
A vector containing the start and end dates of the year-to-date period.
}
\description{
Calculates the start and end dates of the year-to-date period
based on the maximum date in the provided dates. Both partial-
and full-month methods are supported, see \code{rollback_partial_month}.
}
\examples{
ytd_bounds(c("2023-02-04", "2024-02-02", "2024-02-14"))

ytd_bounds(
  c("2023-12-04", "2024-02-14"),
  rollback_partial_month = TRUE
)
}
