% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts_by_status.R
\name{active_accounts_in_range}
\alias{active_accounts_in_range}
\title{List active accounts in a date range}
\usage{
active_accounts_in_range(account_id, order_date, start_date, end_date)
}
\arguments{
\item{account_id}{A vector of account IDs}

\item{order_date}{A vector of order dates corresponding to the account IDs}

\item{start_date}{The start date of the range (inclusive)}

\item{end_date}{The end date of the range (inclusive)}
}
\value{
A vector of unique account IDs that were active in the specified date range
}
\description{
List active accounts in a date range
}
