% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_to_bizday.R
\name{adjust_to_bizday}
\alias{adjust_to_bizday}
\title{Adjust any non-working days to the following business day in a given calendar}
\usage{
adjust_to_bizday(date, calendar)
}
\arguments{
\item{date}{A vector of dates (Date object or coercible with \code{\link[=as.Date]{as.Date()}}).}

\item{calendar}{(character) A QuantLib calendar id (the vector \link[qlcal:calendars]{qlcal::calendars} lists all valid options).}
}
\value{
A vector of Date objects adjusted to the following business day in the specified calendar.
}
\description{
Adjust any non-working days to the following business day in a given calendar
}
\examples{
adjust_to_bizday(c("2025-07-04", "2025-05-31"), "UnitedStates")
adjust_to_bizday(c("2025-01-15", "2025-01-16"), "UnitedStates")
}
\seealso{
\code{\link[qlcal:calendars]{qlcal::calendars()}}, \code{\link[qlcal:adjust]{qlcal::adjust()}}
}
