% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{detect_hidden_logicals}
\alias{detect_hidden_logicals}
\title{Find columns that are character or factor but contain logical values}
\usage{
detect_hidden_logicals(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A character vector of column names that are holding logical data but are typed as character or factor
}
\description{
Find columns that are character or factor but contain logical values
}
\details{
Columns are deemed to hold logical data if they contain only the values
c("T", "TRUE", "True", "true", "F", "FALSE", "False", "false", NA). (matching \code{\link[base:logical]{base::as.logical()}})
}
