% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sales}
\alias{example_sales}
\title{Example sales data}
\format{
A data frame with 5339 rows and 2 variables:
\describe{
\item{account_id}{Character. A unique identifier for the customer that placed the order.}
\item{order_date}{Date. The date of the order.}
}
}
\source{
Generated using random sampling for demonstration purposes.
}
\usage{
example_sales
}
\description{
A dataset containing example sales data over the time period "2022-01-01" to
"2024-12-31".
}
\keyword{datasets}
