% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytd_py.R
\name{is_ytd_comparable}
\alias{is_ytd_comparable}
\title{Is a date comparable for year-to-date (YTD) calculations?}
\usage{
is_ytd_comparable(date, end_date)
}
\arguments{
\item{date}{A date object. (coercible using \code{\link[lubridate:as_date]{lubridate::as_date()}})}

\item{end_date}{A reference end-date to compare against. (i.e. the YTD end date)
(coercible using \code{\link[lubridate:as_date]{lubridate::as_date()}})}
}
\value{
\code{TRUE} if the month and day of date is on or before that of end_date,
\code{FALSE} otherwise.
}
\description{
Is the month and day of date  on or before that of end_date?
Useful for creating YTD comparisons using historical data.
}
\note{
\code{datetimes} are coerced to \code{dates}, so the time component is ignored.
}
\examples{
is_ytd_comparable("2023-05-04", "2024-05-31")
}
