% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrutils-package.R
\docType{package}
\name{mcrutils-package}
\alias{mcrutils}
\alias{mcrutils-package}
\title{mcrutils: Miscellaneous Utilities for Data Cleaning, Analysis, and Reporting}
\description{
Convenience functions for data cleaning, analysis and reporting, including creating and visualizing year-to-date and quarterly analyses, and determining customer account status for customer churn analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcaselli/mcrutils}
  \item \url{https://mcaselli.github.io/mcrutils/}
  \item Report bugs at \url{https://github.com/mcaselli/mcrutils/issues}
}

}
\author{
\strong{Maintainer}: Michael Caselli \email{mkc.develops@gmail.com}

}
\keyword{internal}
