% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_bizdays.R
\name{periodic_bizdays}
\alias{periodic_bizdays}
\title{Calculate the Number of Business Days in Months, Quarters, etc}
\usage{
periodic_bizdays(from, to, by = "month", quantlib_calendars = "UnitedStates")
}
\arguments{
\item{from, to}{start and end dates. Date object or something coercible
with \code{\link[=as.Date]{as.Date()}}.}

\item{by}{periodicity, passed to \code{\link[=seq.Date]{seq.Date()}} (e.g., "month", "quarter",
"year"). See \code{\link[=seq.Date]{seq.Date()}} for more options.}

\item{quantlib_calendars}{(character) A vector of QuantLib calendar ids (the
vector \link[qlcal:calendars]{qlcal::calendars} lists all valid options).}
}
\value{
a tibble with columns:
\describe{
\item{calendar}{the QuantLib calendar used}
\item{start}{the start date of the period}
\item{end}{the end date of the period}
\item{business_days}{the number of business days in the period,
according to the calendar}
}
}
\description{
This function calculates the number of business days in each periodic
interval (e.g., month, quarter) between two dates, using specified QuantLib
calendars for holiday definitions.
}
\examples{
periodic_bizdays(
  from = as.Date("2023-01-01"),
  to = as.Date("2023-12-31"),
  by = "month",
  quantlib_calendars = c("UnitedStates", "UnitedKingdom")
)
}
