% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_bizdays.R
\name{periodic_bizdays}
\alias{periodic_bizdays}
\title{Calculate the number of business days in periodic intervals}
\usage{
periodic_bizdays(from, to, by = "month", quantlib_calendars = "UnitedStates")
}
\arguments{
\item{from, to}{start and end dates. Date object or something coercible
with \code{\link[=as.Date]{as.Date()}}.}

\item{by}{periodicity, passed to \code{\link[=seq.Date]{seq.Date()}} (e.g., "month", "quarter",
"year"). See \code{?seq.Date} for more options.}

\item{quantlib_calendars}{= a character vector of RQuantLib calendar names
(see \code{\link[bizdays:other-calendars]{bizdays::load_quantlib_calendars()}}). Don't inlcude the "QuantLib/"
prefix.}
}
\value{
a tibble with columns:
\describe{
\item{calendar}{the QuantLib calendar used}
\item{start}{the start date of the period}
\item{end}{the end date of the period}
\item{business_days}{the number of business days in the period,
according to the calendar}
}
}
\description{
A Convenience wrapper for \code{\link[bizdays:bizdays]{bizdays::bizdays()}}. This function calculates the
number of business days in periodic intervals (e.g., monthly, quarterly)
between two dates, using specified QuantLib calendars for holiday
definitions.
}
\examples{
periodic_bizdays(
  from = as.Date("2023-01-01"),
  to = as.Date("2023-12-31"),
  by = "month",
  quantlib_calendars = c("UnitedStates", "UnitedKingdom")
)
}
